# perf(DataTable): speed up select-all with large tables