Rewrite `setAllSelectedState` to improve performance with large tables. In my test case on my machine with 10000 rows, selecting all rows previously took 14000ms, this fix reduces it to 8ms. It uses one instance of (local) variable mutation.

Closes #6146

The `setAllSelectedState` function is called when you click the *select-all* (or *deselect all*), or select one row then hit the *cancel* button on the bulk-action toolbar. It appears to be "accidentally quadratic" and starts to become an issue with more than a few thousand rows. To avoid it, I do a local mutation of the accumulator rather than re-splatting out the entire object each iteration.

For 10000 rows, this reduces the original 14 second time (this is for Vite transpiled code, 8 seconds for native), down to ~250ms. Most of that time is in the `filteredRowIds.includes` check. So I added an extra flag (`isFiltered`) to avoid calling when the filter is not in use. When there are no filtered rows, the function runs in ~8ms. 

#### Changelog

#### Testing / Reviewing

The fix does not break any existing tests, and should not introduce any changes to the existing behavior of the DataTable: Selecting-all/deselecting all should work as expected. When using the bulk-action toolbar, clicking cancel should deselect the row. If rows are filtered, clicking select-all should select all filtered rows.

To test the performance, you need to make a table with a lot of (at least 10000) rows. With a large dataset like this, the performance improvements are obvious, but if you want to measure it then wrap calls to `setAllSelectedState` with console.time messages. E.g, inside the DataTable's `handleOnCancel` function:
```
...
console.log("Cancel button was hit - call setAllSelectedState");
console.time("setAllSelectedState");
const output = this.setAllSelectedState(state, false, this.getFilteredRowIds());
console.timeEnd("setAllSelectedState");
return {
  shouldShowBatchActions: false,
   ...output
};
```