@mrspeaker 2023-05-24
> 



> Do you think we could speed it up even more by not relying on `reduce` and just using a for loop instead to build up the rowsByID?

@tay1orjones - strangely, I thought the same - but when I tested it the `for` loop was every-so-slightly slower! I'm not sure why (and I bet it's browser specific), but in either case the time for the loop itself is negligible compared to the loop body. I decided to keep the `reduce` there to more closely match the coding style of the code base.